pyproject.toml
[project]
name = "end-to-end-medical-chatbot"
version = "0.1.0"
description = "GenAI chatbot for medical Q&A"
authors = [
    {name = "Boyang Wan",email = "boyangwan12@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "flask (>=3.1.1,<4.0.0)",
    "pypdf (>=5.5.0,<6.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pinecone (>=7.0.1,<8.0.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "langchain-openai (>=0.3.18,<0.4.0)",
    "langchain-experimental (>=0.3.4,<0.4.0)",
    "huggingface-hub (>=0.20.0)",
    "protobuf (>=6.31.0,<7.0.0)",
    "pyreadline3 (>=3.5.4,<4.0.0)",

]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

